Move Featurizer into util package
1. Moved Featurizer related class into util package
2. Add publish gradle task back
3. Removes optional dependencies
