+++ extensions/tablesaw/README.md
@@ -23,6 +23,6 @@ You can pull the module from the central Maven repository by including the follo
 <dependency>
     <groupId>ai.djl.tablesaw</groupId>
     <artifactId>tablesaw</artifactId>
-    <version>0.17.0</version>
+    <version>0.18.0-SNAPSHOT</version>
 </dependency>
 ```
