+++ gradle.properties
@@ -33,6 +33,7 @@ awssdk_version=2.17.151
 hadoop_version=3.3.1
 javacpp_version=1.5.6
 protobuf_version=3.19.2
+tablesaw_version=0.43.1
 
 antlr_version=4.9.3
 testng_version=7.5
